# openapi.yaml
openapi: 3.0.3
info:
  title: Kick Accounting API
  version: 1.0.0
  description: >
    Backend accounting platform with General Ledger, journal entries,
    and financial reports (Trial Balance, P&L, Balance Sheet).
servers:
  - url: /api
    description: API base path
paths:
...
  /reports/trial-balance:
    get:
      tags: [Reports]
      summary: Trial Balance
      operationId: getTrialBalance
      description: >
        Summarizes all accounts with total debits, credits, and net balance
        for a date range. Grand total debits must equal grand total credits.
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Period start (YYYY-MM-DD)
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Period end (YYYY-MM-DD)
      responses:
        "200":
          description: Trial Balance report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TrialBalance"
              example:
                data:
                  startDate: "2025-01-01"
                  endDate: "2025-12-31"
                  rows:
                    - code: "10100"
                      name: "Bank Account - Chase Business Checking - 0205"
                      class: Assets
                      type: Bank Account
                      totalDebit: 248500.90
                      totalCredit: 133245.73
                      balance: 115255.17
                    - code: "40000"
                      name: Revenue
                      class: Income
                      type: Revenue
                      totalDebit: 0
                      totalCredit: 278121.57
                      balance: 278121.57
                  grandTotalDebit: 556742.14
                  grandTotalCredit: 556742.14
                  isBalanced: true
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "startDate: startDate must be YYYY-MM-DD"
  /reports/profit-and-loss:
    get:
      tags: [Reports]
      summary: Profit & Loss (Income Statement)
      operationId: getProfitAndLoss
      description: >
        Revenue minus expenses over a period.
        Sections: Income, COGS, Gross Profit, Operating Expenses,
        Net Operating Income, Other Income/Expenses, Net Income.
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: P&L report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ProfitAndLoss"
              example:
                data:
                  startDate: "2025-01-01"
                  endDate: "2025-12-31"
                  income:
                    label: Income
                    accounts:
                      - code: "40000"
                        name: Revenue
                        class: Income
                        type: Revenue
                        totalDebit: 0
                        totalCredit: 278121.57
                        balance: 278121.57
                    total: 278121.57
                  costOfGoodsSold:
                    label: Cost of Goods Sold
                    accounts: []
                    total: 0
                  grossProfit: 278121.57
                  operatingExpenses:
                    label: Operating Expenses
                    accounts:
                      - code: "60000"
                        name: "Salaries and Wages"
                        class: Expenses
                        type: "Salaries and Wages"
                        totalDebit: 45833.37
                        totalCredit: 0
                        balance: 45833.37
                    total: 152765.51
                  netOperatingIncome: 125356.06
                  otherIncomeExpenses:
                    label: Other Income/Expenses
                    accounts:
                      - code: "80100"
                        name: "Interest Income"
                        class: Expenses
                        type: "Interest Income"
                        totalDebit: 0
                        totalCredit: 536.00
                        balance: -536.00
                    total: -7338.00
                  netIncome: 132694.06
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /reports/balance-sheet:
    get:
      tags: [Reports]
      summary: Balance Sheet
      operationId: getBalanceSheet
      description: >
        Financial position at a point in time.
        Assets = Liabilities + Equity.
        Net Income from current fiscal year P&L (Jan 1 of asOfDate year through asOfDate) is injected into Equity.
      parameters:
        - name: asOfDate
          in: query
          required: true
          schema:
            type: string
            format: date
          description: Report date (YYYY-MM-DD)
      responses:
        "200":
          description: Balance Sheet report
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/BalanceSheet"
              example:
                data:
                  asOfDate: "2025-12-31"
                  currentAssets:
                    label: Current Assets
                    accounts:
                      - code: "10100"
                        name: "Bank Account - Chase Business Checking - 0205"
                        class: Assets
                        type: Bank Account
                        totalDebit: 248500.90
                        totalCredit: 133245.73
                        balance: 115255.17
                    total: 85917.19
                  nonCurrentAssets:
                    label: Non-Current Assets
                    accounts: []
                    total: 0
                  totalAssets: 85917.19
                  currentLiabilities:
                    label: Current Liabilities
                    accounts:
                      - code: "21000"
                        name: "Credit Card - Amex Business Gold Card - 1001"
                        class: Liabilities
                        type: Credit Card
                        totalDebit: 10520.00
                        totalCredit: 20298.62
                        balance: 9778.62
                    total: 19395.13
                  nonCurrentLiabilities:
                    label: Non-Current Liabilities
                    accounts: []
                    total: 0
                  totalLiabilities: 19395.13
                  equity:
                    label: Equity
                    accounts:
                      - code: "31000"
                        name: "Owner Draws"
                        class: Equity
                        type: "Owner Draws"
                        totalDebit: 51000.00
                        totalCredit: 0
                        balance: -51000.00
                      - code: NET-INCOME
                        name: "Net Income (Current Period)"
                        class: Equity
                        type: Retained Earnings
                        totalDebit: 0
                        totalCredit: 117522.06
                        balance: 117522.06
                    total: 66522.06
                  totalLiabilitiesAndEquity: 85917.19
                  isBalanced: true
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "asOfDate: asOfDate must be YYYY-MM-DD"
...
components:
  schemas:
    AccountClass:
      type: string
      enum:
        - Assets
        - Liabilities
        - Equity
        - Income
        - Expenses
...
    TrialBalanceRow:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        class:
          $ref: "#/components/schemas/AccountClass"
        type:
          type: string
        totalDebit:
          type: number
        totalCredit:
          type: number
        balance:
          type: number
    TrialBalance:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        rows:
          type: array
          items:
            $ref: "#/components/schemas/TrialBalanceRow"
        grandTotalDebit:
          type: number
        grandTotalCredit:
          type: number
        isBalanced:
          type: boolean
    ReportSection:
      type: object
      properties:
        label:
          type: string
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/TrialBalanceRow"
        total:
          type: number
    ProfitAndLoss:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        income:
          $ref: "#/components/schemas/ReportSection"
        costOfGoodsSold:
          $ref: "#/components/schemas/ReportSection"
        grossProfit:
          type: number
        operatingExpenses:
          $ref: "#/components/schemas/ReportSection"
        netOperatingIncome:
          type: number
        otherIncomeExpenses:
          $ref: "#/components/schemas/ReportSection"
        netIncome:
          type: number
    BalanceSheet:
      type: object
      properties:
        asOfDate:
          type: string
          format: date
        currentAssets:
          $ref: "#/components/schemas/ReportSection"
        nonCurrentAssets:
          $ref: "#/components/schemas/ReportSection"
        totalAssets:
          type: number
        currentLiabilities:
          $ref: "#/components/schemas/ReportSection"
        nonCurrentLiabilities:
          $ref: "#/components/schemas/ReportSection"
        totalLiabilities:
          type: number
        equity:
          $ref: "#/components/schemas/ReportSection"
        totalLiabilitiesAndEquity:
          type: number
        isBalanced:
          type: boolean
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message